server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "基于工作量证明的验证码系统"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.11",
    "pyjwt>=2.10.1",
    "redis>=5.2.1",
    "uvicorn>=0.34.0",
]
